+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
-# v3.2.0 (started on January 1st, 2026):
-* [Feature] Support district signals for stock value and capacity.
+# v3.2.1 (started on January 1st, 2026):
+* [Feature] Support district signals for stock value and capacity: `District.ResourceStock.<GoodId>` and `District.ResourceCapacity.<GoodId>`.
+* [Feature] Add change recipe action: `Manufactory.SetRecipe(<RecipeId>)`.
 * [Change] Deprecate debug signal: `Debug.DistrictStockTracker.<*>`.
 * [Change] Drop tool "Prioritize by haulers on finish building". This option can now be set on unfinished building via the stock game UI.
 * [Change] Deprecate operators `getnum` and `getstr`. Now, you should use: `getvalue`, `getelement`, and `getlen`. Good news: the value type is now detected from the property.